[MPEG2d] Align code with open-source
Issue: -
Test: mpeg2d_conf

+++ _studio/shared/umc/codec/mpeg2_dec/hw/src/umc_mpeg2_bitstream.cpp
@@ -296,8 +296,8 @@ namespace UMC_MPEG2_DECODER
         if (CheckBSLeft())
             throw mpeg2_exception(UMC::UMC_ERR_INVALID_STREAM);
 
-        uint32_t width_in_MBs = (uint32_t)(mfx::align2_value<uint32_t>(slice.GetSeqHeader().horizontal_size_value, 16) / 16);
-        uint32_t height_in_MBs = (uint32_t)(mfx::align2_value<uint32_t>(slice.GetSeqHeader().vertical_size_value, 16) / 16);
+        uint32_t width_in_MBs  = mfx::align2_value<uint32_t>(slice.GetSeqHeader().horizontal_size_value, 16) / 16u;
+        uint32_t height_in_MBs = mfx::align2_value<uint32_t>(slice.GetSeqHeader().vertical_size_value, 16) / 16u;
 
         // invalid slice - skipping it
         if (slice.sliceHeader.slice_vertical_position > height_in_MBs)
